Rework the targets to work on first build.
This is a great solution, however with a few minor changes it will be
even better :).

If we move the `ItemGroup` inside the `Target` and after the call to
`Csc` it will add the assembly after it has been compiled. The good
thing is that even if the `Target` is skipped, msbuild will still
evaluate the `ItemGroup`. This means the assembly will always get added.

Because we are doing the `ItemGroup` changes late in the compilation process
we should use the `ReferencePath` rather than `References` `ItemGroup`.
This will make sure it gets passed tot the fsharp compiler.

One last change is to add the assembly to the `FileWrites` group. This
will prevent it from being deleted between builds.